Silence a few compiler warnings by explicit casting
These may not be the best long-term fixes for these warnings, so please think
before forward-porting these to the master branch.

+++ dune/foamgrid/foamgrid/foamgrid.cc
@@ -279,7 +279,7 @@ bool Dune::FoamGrid<dimgrid, dimworld>::adapt()
       // check if we have same level elements, if so this facet stays
       bool hasSameLevelElements=false;
       for (ElementIter elem=facet->elements_.begin(); elem != facet->elements_.end(); ++elem)
-        hasSameLevelElements = hasSameLevelElements || (*elem)->level()==level;
+        hasSameLevelElements = hasSameLevelElements || (*elem)->level()==(size_t)level;
 
       assert(facet->willVanish_!=hasSameLevelElements);
